requirements.txt
-e svn+http://code.djangoproject.com/svn/django/trunk/#egg=django
django-nose
http://nodeload.github.com/django-debug-toolbar/django-debug-toolbar/tarball/master#egg=django-debug-toolbar
http://nodeload.github.com/jessemiller/HamlPy/tarball/master#egg=HamlPy
http://nodeload.github.com/chartjes/djaml/tarball/master#egg=djaml
mechanize